monitor: separate monitor and listener code
This is a simple move to reflect that these are different things.
